refactor(profile): replace the hand-rolled tab bar with routed hub tabs
Profile was the last page running its own signal-driven tab bar, which meant
its sections were not deep-linkable, did not respond to back/forward, and all
loaded eagerly. It also forced a special case in the dashboard: hubTabs()
suppressed the shared bar for the profile hub, and hubs.const.ts carried a
sentinel tab that existed only so hubForUrl could resolve the URL.

Split the page into three lazily-loaded child routes — CV, Personal details,
Cover letters — and fill in the real profile hub. Nearly all of the old
component's state was CV-specific and moves wholesale into the CV tab; the
shell keeps only the header and the one-time bootstrap. Because the tabs are
now separate routed components they cannot read a loading flag off a parent,
so ProfileService owns a `loaded` signal that each tab derives its spinner
from — no tab refetches on navigation.

The stylesheet is split the same way, with the genuinely shared rules (loading
state, buttons, alerts, details-card, empty state) in a partial each tab uses.
The .page-tabs/.page-tab/.tab-panel rules are deleted with the markup they
styled.
