+++ frontend/src/app/pages/profile/profile.component.spec.ts
@@ -4,29 +4,13 @@ import { provideRouter } from '@angular/router';
 import { of, throwError } from 'rxjs';
 import { ProfileComponent } from './profile.component';
 import { ProfileService } from '../../core/services/profile.service';
-import { ApplicationsService } from '../../core/services/applications.service';
-import { CoverLetterTemplatesService } from '../../core/services/cover-letter-templates.service';
-import { TranslateResponse } from './models/translate-response.model';
+import { makeProfile } from './testing/make-profile';
 
-function makeProfile(over: Partial<Record<string, unknown>> = {}) {
-  return {
-    id: 'p-en',
-    name: 'Ada Lovelace',
-    email: null,
-    phone: null,
-    location: null,
-    sections: [],
-    raw_tex: '',
-    language: 'en',
-    skills: ['python'],
-    linkedin_url: null,
-    github_url: null,
-    portfolio_url: null,
-    ...over,
-  };
-}
-
-describe('ProfileComponent', () => {
+/**
+ * The shell only owns the one-time bootstrap; the tabs are routed children and
+ * are covered by their own specs.
+ */
+describe('ProfileComponent (shell)', () => {
   function mount(
     opts: {
       profiles?: Record<string, unknown>;
@@ -34,198 +18,57 @@ describe('ProfileComponent', () => {
       getCurrentProfile?: () => unknown;
     } = {},
   ) {
-    const profiles = signal<Record<string, unknown>>(opts.profiles ?? {});
-    const activeLanguage = signal('en');
     const profileService = {
-      profiles,
-      activeLanguage,
+      profiles: signal<Record<string, unknown>>(opts.profiles ?? {}),
+      activeLanguage: signal('en'),
       profile: signal(opts.profiles?.['en'] ?? null),
+      loaded: signal(false),
       listProfiles: opts.listProfiles ?? (() => of(Object.values(opts.profiles ?? {}))),
       getCurrentProfile: opts.getCurrentProfile ?? (() => of(makeProfile())),
-      uploadFile: () => of(makeProfile()),
-      uploadCV: () => of(makeProfile()),
-      translate: vi.fn(() =>
-        of({
-          profile: makeProfile({ id: 'p-es', language: 'es', machine_translated: true }),
-          pdf_ok: true,
-          compile_error: null,
-        } as unknown as TranslateResponse),
-      ),
-      downloadCvPdf: vi.fn(() => of(new Blob(['%PDF'], { type: 'application/pdf' }))),
     };
 
     TestBed.configureTestingModule({
       imports: [ProfileComponent],
-      providers: [
-        provideRouter([]),
-        { provide: ProfileService, useValue: profileService },
-        { provide: ApplicationsService, useValue: { listCoverLetters: () => of([]) } },
-        { provide: CoverLetterTemplatesService, useValue: { list: () => of([]) } },
-      ],
+      providers: [provideRouter([]), { provide: ProfileService, useValue: profileService }],
     });
     const fixture = TestBed.createComponent(ProfileComponent);
     fixture.detectChanges();
     return { fixture, profileService };
   }
 
-  it('clears initialLoading after listProfiles succeeds (happy path)', () => {
+  it('marks the profile loaded after listProfiles succeeds (happy path)', () => {
     const listProfiles = vi.fn(() => of([makeProfile()]));
-    const { fixture } = mount({ profiles: { en: makeProfile() }, listProfiles });
-    const cmp = fixture.componentInstance;
+    const { profileService } = mount({ profiles: { en: makeProfile() }, listProfiles });
 
     expect(listProfiles).toHaveBeenCalled();
-    expect(cmp.initialLoading()).toBe(false);
-    expect(cmp.uploadedLanguages()).toEqual(['en']);
+    expect(profileService.loaded()).toBe(true);
   });
 
-  it('falls back to getCurrentProfile then clears initialLoading when listProfiles fails', () => {
+  it('falls back to getCurrentProfile then marks loaded when listProfiles fails', () => {
     const getCurrentProfile = vi.fn(() => of(makeProfile()));
-    const { fixture } = mount({
+    const { profileService } = mount({
       listProfiles: () => throwError(() => new Error('boom')),
       getCurrentProfile,
     });
-    const cmp = fixture.componentInstance;
 
     expect(getCurrentProfile).toHaveBeenCalled();
-    expect(cmp.initialLoading()).toBe(false);
+    expect(profileService.loaded()).toBe(true);
   });
 
-  it('clears initialLoading even when both fetches fail (error state)', () => {
-    const { fixture } = mount({
+  it('marks loaded even when both fetches fail, so tabs stop spinning (error state)', () => {
+    const { profileService } = mount({
       listProfiles: () => throwError(() => new Error('boom')),
       getCurrentProfile: () => throwError(() => new Error('boom2')),
     });
 
-    expect(fixture.componentInstance.initialLoading()).toBe(false);
+    expect(profileService.loaded()).toBe(true);
   });
 
-  it('surfaces an upload error and clears loading when uploadFile fails', () => {
-    const profiles = signal<Record<string, unknown>>({});
-    const profileService = {
-      profiles,
-      activeLanguage: signal('en'),
-      profile: signal(null),
-      listProfiles: () => of([]),
-      getCurrentProfile: () => of(makeProfile()),
-      uploadFile: () => throwError(() => ({ error: { detail: 'bad file' } })),
-      uploadCV: () => of(makeProfile()),
-    };
-    TestBed.configureTestingModule({
-      imports: [ProfileComponent],
-      providers: [
-        provideRouter([]),
-        { provide: ProfileService, useValue: profileService },
-        { provide: ApplicationsService, useValue: { listCoverLetters: () => of([]) } },
-        { provide: CoverLetterTemplatesService, useValue: { list: () => of([]) } },
-      ],
-    });
-    const fixture = TestBed.createComponent(ProfileComponent);
-    fixture.detectChanges();
-    const cmp = fixture.componentInstance;
-
-    cmp.selectedFile.set(new File(['x'], 'cv.pdf', { type: 'application/pdf' }));
-    cmp.uploadFile();
-
-    expect(cmp.error()).toBe('bad file');
-    expect(cmp.loading()).toBe(false);
-  });
-
-  it('replaceCv pre-selects the active language and opens the upload form', () => {
-    const { fixture, profileService } = mount({ profiles: { en: makeProfile() } });
-    const cmp = fixture.componentInstance;
-    (profileService.activeLanguage as { set: (v: string) => void }).set('en');
-
-    cmp.replaceCv();
-
-    expect(cmp.showUploadForm()).toBe(true);
-    expect(cmp.uploadIntent()).toBe('replace');
-    expect(cmp.language()).toBe('en');
-    expect(cmp.selectedFile()).toBeNull();
-  });
-
-  it('hides the personal-details form until Edit is clicked, then swaps back on save/cancel', () => {
+  it('renders the routed tab outlet rather than a hand-rolled tab bar', () => {
     const { fixture } = mount({ profiles: { en: makeProfile() } });
-    const cmp = fixture.componentInstance;
-    cmp.pageTab.set('personal');
-    fixture.detectChanges();
     const el = fixture.nativeElement as HTMLElement;
 
-    // read-only card visible, form hidden by default
-    expect(cmp.editingPersonal()).toBe(false);
-    expect(el.querySelector('.details-grid')).not.toBeNull();
-    expect(el.querySelector('app-manual-fields-form')).toBeNull();
-
-    // clicking Edit reveals the form and hides the read-only grid
-    const editBtn = [...el.querySelectorAll('button')].find(
-      (b) => b.textContent?.trim() === 'Edit',
-    )!;
-    editBtn.click();
-    fixture.detectChanges();
-    expect(cmp.editingPersonal()).toBe(true);
-    expect(el.querySelector('app-manual-fields-form')).not.toBeNull();
-    expect(el.querySelector('.details-grid')).toBeNull();
-
-    // saved/cancelled return to the read-only view
-    cmp.editingPersonal.set(false);
-    fixture.detectChanges();
-    expect(el.querySelector('app-manual-fields-form')).toBeNull();
-    expect(el.querySelector('.details-grid')).not.toBeNull();
-  });
-
-  it('shows a profile setup guide in the personal details tab', () => {
-    const { fixture } = mount({ profiles: { en: makeProfile() } });
-    const cmp = fixture.componentInstance;
-    cmp.pageTab.set('personal');
-    fixture.detectChanges();
-
-    expect(fixture.nativeElement.textContent).toContain('Make your profile ready to use');
-  });
-
-  it('addAnotherLanguage marks the intent as add', () => {
-    const { fixture } = mount({ profiles: { en: makeProfile() } });
-    const cmp = fixture.componentInstance;
-
-    cmp.addAnotherLanguage();
-
-    expect(cmp.showUploadForm()).toBe(true);
-    expect(cmp.uploadIntent()).toBe('add');
-    expect(cmp.language()).toBe('es');
-  });
-
-  it('calls translate with the other language', () => {
-    const { fixture, profileService } = mount({ profiles: { en: makeProfile() } });
-    const cmp = fixture.componentInstance;
-
-    cmp.translateToOther();
-
-    expect(profileService.translate).toHaveBeenCalledWith('es');
-  });
-
-  it('surfaces the compiler error when the translated PDF does not compile', () => {
-    const { fixture, profileService } = mount({ profiles: { en: makeProfile() } });
-    const cmp = fixture.componentInstance;
-    profileService.translate = vi.fn(() =>
-      of({
-        profile: makeProfile({ id: 'p-es', language: 'es', machine_translated: true }),
-        pdf_ok: false,
-        compile_error: 'xelatex exited with code 1\nMissing \\begin{document}.',
-      } as unknown as TranslateResponse),
-    );
-
-    cmp.translateToOther();
-
-    expect(cmp.translateWarning()).not.toBe('');
-    expect(cmp.translateCompileError()).toContain('Missing \\begin{document}');
-  });
-
-  it('downloads the PDF blob for a language', () => {
-    const clickSpy = vi.spyOn(HTMLAnchorElement.prototype, 'click').mockImplementation(() => {});
-    const { fixture, profileService } = mount({ profiles: { en: makeProfile() } });
-    const cmp = fixture.componentInstance;
-
-    cmp.downloadPdf('en');
-
-    expect(profileService.downloadCvPdf).toHaveBeenCalledWith('en');
-    clickSpy.mockRestore();
+    expect(el.querySelector('router-outlet')).not.toBeNull();
+    expect(el.querySelector('.page-tabs')).toBeNull();
   });
 });
