+++ frontend/src/app/pages/profile/tabs/cover-letters-tab/profile-cover-letters-tab.component.scss
@@ -1,0 +1,1 @@
+@use '../profile-tab-shared' as *;
